requirements.txt
APScheduler==3.9.1
asgiref==3.5.2
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
defusedxml==0.7.1
Django==3.2
django-appconf==1.0.5
django-crispy-forms==1.14.0
django-cryptography==1.1
django-taggit==3.0.0
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.3
numpy==1.23.0
oauthlib==3.2.0
pycparser==2.21
PyJWT==2.4.0
python3-openid==3.2.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
twython==3.9.1
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
whitenoise==6.2.0
dj-database-url==0.5.0
psycopg2==2.9.3
python-decouple==3.6
whitenoise==6.2.0